Wrap cls stuff into cls.js initializer stex.cls
