+++ lib/initializers/errors.js
@@ -2,7 +2,6 @@ var Initializer = require("../initializer");
 var raven       = require("raven");
 var _           = require("lodash");
 var domain      = require('domain');
-var getNamespace = require('continuation-local-storage').getNamespace;
 
 
 Initializer.add('startup', 'stex.errors', ['stex.initialize-app'], function(stex) {
@@ -28,8 +27,7 @@ Initializer.add('startup', 'stex.errors', ['stex.initialize-app'], function(stex
 
     if (stex.sentry) {
       // put requestId on err
-      var namespace = getNamespace('org.stellar.stex');
-      err.requestId = namespace.get("requestId");
+      err.requestId = stex.cls.get("requestId");
       stex.sentry.captureError(err);
     }
   }
