+++ lib/initializers/express.js
@@ -1,9 +1,4 @@
 var Initializer = require("../initializer");
-var uuid = require('uuid');
-var createNamespace = require('continuation-local-storage').createNamespace,
-    namespace = createNamespace('org.stellar.stex');
-var patchBluebird = require("cls-bluebird");
-patchBluebird(namespace);
 
 Initializer.add('startup', 'stex.express', ['stex.config', 'stex.logging', 'stex.db', 'stex.newrelic'], function(stex) {
   var express    = require("express");
@@ -63,24 +58,5 @@ Initializer.add('startup', 'stex.express', ['stex.config', 'stex.logging', 'stex
     });
 
     parser(req, res, next);
-  });
-
-  // create a transaction id for each request
-  app.use(function(req, res, next) {
-    var requestId = uuid.v4();
-
-    // wrap the events from request and response
-    namespace.bindEmitter(req);
-    namespace.bindEmitter(res);
-
-    // run following middleware in the scope of
-    // the namespace we created
-    namespace.run(function() {
-
-      // set requestId on the namespace, makes it
-      // available for all continuations
-      namespace.set('requestId', requestId);
-      next();
-    });
   });
 });
