+++ lib/initializers/logging.js
@@ -1,6 +1,5 @@
-var Initializer = require("../initializer");
-var util        = require("util");
-var getNamespace = require('continuation-local-storage').getNamespace;
+var Initializer     = require("../initializer");
+var util            = require("util");
 
 /**
  * This logging subsystem is setup to format any logging data as json.  By default,
@@ -16,8 +15,7 @@ Initializer.add('startup', 'stex.logging', ['stex.config'], function(stex) {
   stex.log.addTransform(function (element) {
     // check this is a meta data element
     if (element.file) {
-      var namespace = getNamespace('org.stellar.stex');
-      element.requestId = namespace.get('requestId');
+      element.requestId = stex.cls.get('requestId');
       return element;
     }
     return null;
